# Improve 128-bit hash performance